// kraken_frontend/src/api/generated/models/FindingDefinitionUsage.ts
export interface FindingDefinitionUsage {
    /**
     * The uuid of the finding
     * @type {string}
     * @memberof FindingDefinitionUsage
     */
    uuid: string;
    /**
     * 
     * @type {FindingSeverity}
     * @memberof FindingDefinitionUsage
     */
    severity: FindingSeverity;
    /**
     * The point in time the finding was created
     * @type {Date}
     * @memberof FindingDefinitionUsage
     */
    createdAt: Date;
    /**
     * 
     * @type {SimpleWorkspace}
     * @memberof FindingDefinitionUsage
     */
    workspace: SimpleWorkspace;
    /**
     * The number of affected domains
     * @type {number}
     * @memberof FindingDefinitionUsage
     */
    affectedDomains: number;
    /**
     * The number of affected hosts
     * @type {number}
     * @memberof FindingDefinitionUsage
     */
    affectedHosts: number;
    /**
     * The number of affected ports
     * @type {number}
     * @memberof FindingDefinitionUsage
     */
    affectedPorts: number;
    /**
     * The number of affected services
     * @type {number}
     * @memberof FindingDefinitionUsage
     */
    affectedServices: number;
}